feat: add database seeder with realistic auto parts data

+++ StockDemo.Api/Models/Product.cs
@@ -2,13 +2,13 @@
 
 public class Product
 {
-    public int Id {get;set;}
-    public string Reference {get;set;}
-    public string Label {get;set;}
-    public string Brand {get;set;}
-    public string? Description {get;set;}
-    public decimal? Price {get;set;} 
-    public DateTime CreatedAt {get;set;}
-    public DateTime UpdatedAt {get;set;}
+    public int Id { get; set; }
+    public string Reference { get; set; }
+    public string Label { get; set; }
+    public string Brand { get; set; }
+    public string? Description { get; set; }
+    public double? Price { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public DateTime UpdatedAt { get; set; }
 }
